create an abstract base version of readyonlyconfigurationservice so it can be extended for testing
